transition to level 3 semi working
Discovered huge problems
removeFade function is repeatedly called, though out the game but
endFade does not get repeatedly called. Just removeFade is repeatedly
called.

Second the fade between level 2 and 3 i needs working. The sound plays
and the screen fades but it fades back into the report screen rather
than the next level.

The report ding (played right when the report is displayed) also needs
to be placed. Need the removeFade bug fixed for it to work.